Search tweaks: move to use extension, commands, add search to SuperDoc.js
