+++ packages/super-editor/src/core/Editor.js
@@ -1,7 +1,6 @@
 import { EditorState } from 'prosemirror-state';
 import { EditorView } from 'prosemirror-view';
 import { DOMParser, DOMSerializer } from 'prosemirror-model';
-import { search, SearchQuery, setSearchState, getMatchHighlights } from 'prosemirror-search';
 import { yXmlFragmentToProseMirrorRootNode } from 'y-prosemirror';
 import { helpers } from '@core/index.js';
 import { EventEmitter } from './EventEmitter.js';
@@ -230,30 +229,6 @@ export class Editor extends EventEmitter {
   #onFocus({ editor, event }) {
     this.toolbar?.setActiveEditor(editor);
     this.options.onFocus({ editor, event });
-  }
-
-  goToFirstMatch() {
-    const highlights = getMatchHighlights(this.view.state);
-    if (!highlights || !highlights.children?.length) return;
-    
-    const match = highlights.children.find(item => item.local);
-    const firstSearchItemPosition = highlights.children[0] + match.local[0].from + 1;
-    this.view.domAtPos(firstSearchItemPosition)?.node?.scrollIntoView(true);
-  }
-  
-  doSearch(text) {
-    const query = new SearchQuery({
-      search: text,
-      caseSensitive: false,
-      regexp: false,
-      wholeWord: false 
-    });
-    const tr = this.view.state.tr;
-    setSearchState(tr, query);
-    this.view.dispatch(tr);
-
-    this.goToFirstMatch();
-    return getMatchHighlights(this.view.state);
   }
 
   setToolbar(toolbar) {
@@ -670,13 +645,8 @@ export class Editor extends EventEmitter {
       state: EditorState.create(state),
     });
     
-    const searchPlugin = search();
-    
     const newState = this.state.reconfigure({
-      plugins: [
-        ...this.extensionService.plugins,
-        searchPlugin,
-      ],
+      plugins: [...this.extensionService.plugins],
     });
 
     this.view.updateState(newState);
