+++ packages/super-editor/src/extensions/index.js
@@ -56,6 +56,7 @@ import { PopoverPlugin } from './popover-plugin/index.js';
 import { TrackChanges } from './track-changes/index.js';
 import { Pagination } from './pagination/index.js';
 import { LinkedStyles } from './linked-styles/linked-styles.js';
+import { Search } from './search/index.js';
 
 // Helpers
 import { trackChangesHelpers } from './track-changes/index.js';
@@ -161,6 +162,7 @@ const getStarterExtensions = () => {
     ShapeContainer,
     ShapeTextbox,
     ContentBlock,
+    Search,
   ];
 };
 
@@ -220,4 +222,5 @@ export {
   AiMark,
   AiAnimationMark,
   AiPlugin,
+  Search,
 };
