+++ packages/super-editor/src/extensions/search/index.js
@@ -1,0 +1,1 @@
+export * from './search.js';
